requirements.txt

discord.py>=2.5.2
pynacl>=1.5.0
python-dotenv>=1.1.0
spotipy>=2.25.1
yt-dlp>=2025.3.31
ffmpeg-python>=0.2.0
ffmpeg-python
fake-useragent
